dynamic_memory_go/python/requirements.txt
grpcio>=1.62.0
protobuf>=4.21.0
numpy>=1.24.0
